+++ src/Calculate/Calculate.jl
@@ -17,6 +17,8 @@ using XCALibre.Fields
 using XCALibre.ModelFramework
 using XCALibre.Discretise
 using XCALibre.Solve
+using XCALibre.Simulate
+
 
 include("Calculate_0_gradient.jl")
 include("Calculate_0_gradient_cell_limiter.jl")