+++ src/ModelPhysics/Turbulence/Turbulence.jl
@@ -22,6 +22,7 @@ include("RANS_functions.jl")
 include("RANS_tensor_algebra.jl")
 include("RANS_laminar.jl")
 include("RANS_kOmega.jl")
+include("RANS_kOmegaSST.jl")
 include("RANS_kOmegaLKE.jl")
 
 # LES models