+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 *  `SetFields` utility that allows to set a field to desired value within a box / circle / sphere [#73](@ref)
 *  Helmholtz Energy equation of state and supporting framework for it for H2 and N2 [#75](@ref)
 *  Added `RotatingWall` velocity boundary condition [#81](@ref)
+*  Added `KOmegaSST` turbulence model [#82](@ref)
 
 
 ### Fixed